Added song clas to handle requests, fixed error messages, and moved __init.py to GoogleEchoMusicService.py.
